core: p3m: Split off halo code
