+++ src/core/electrostatics_magnetostatics/p3m-dipolar.hpp
@@ -41,6 +41,7 @@
 #ifdef DP3M
 #include "electrostatics_magnetostatics/dipole.hpp"
 #include "fft.hpp"
+#include "p3m-halo.hpp"
 #include "p3m-common.hpp"
 #include "particle_data.hpp"
